// typescript/user-interface/packages/weavess-core/src/ts/types/types.ts
export interface TimeRange {
  /** Start Time in seconds */
  startTimeSecs: number;

  /** End Time in seconds */
  endTimeSecs: number;
}
// ...
export interface DataBySampleRate {
  /** Epoch start time in seconds */
  startTimeSecs: number;

  /** Epoch end time in seconds */
  endTimeSecs: number;

  /** Sample Rate */
  sampleRate: number;

  /** Collection representing data segment data */
  values: Float32Array | number[];
}
// ...
export interface TimeValuePair {
  timeSecs: number;
  value: number;
}